Allow trusted proxy to fall back to spnego
In some scenarios we would also like to allow service users to be
authenticated and perform a Cruise Control action that they're
authorized to do. A nice option to enable this is to allow trusted
proxy to fall back to spnego.
That is when this functionality is enabled via the newly created
trusted.proxy.spnego.fallback.enabled option and no doAs user provided,
the TrustedProxySecurityProvider will authenticate and authorize the
service user.

This fixes #1617